### 2021-05-10 #443 @makinin
rdbhapi : R package to simplify API access to data on higher education in Norway
Submitting Author Name: Marija Ninic
Submitting Author Github Handle: @makinin
Repository:   https://github.com/makinin/rdbhapi
Version submitted: 2.1.0
Submission type: Standard

Version accepted: TBD

```
Package: rdbhapi
Title: Interface to 'DBH-API'
Version: 2.1.0
Authors@R: c(person(given = "Marija",family = "Ninic",role = c("cre", "aut"),email = "marija.ninic@nsd.no", comment=c(ORCID ="0000-0001-8484-5817")),person(given=c("Sigve", "Hofland"), family="Berge", role = "aut"))
Description: Downloads 'DBH-API'<https://dbh.nsd.uib.no/tjenester.action> data as dataframes. These data are in machine-readable format and provides information about education, finances, research and publications.
    httr,
    rjson,
    stats

```

	- [ x] data retrieval

- Explain how and why the package falls under these categories (briefly, 1-2 sentences): his package falls under the category of data retrieval because it simplify API access from to data on higher education in Norway, as provided by NSD - Norwegian Centre for Research Data. This package read data from API into R dataframe.

-   Who is the target audience and what are scientific applications of this package? Students of social sciences and researches who work with data about higher education in Norway.

-   Are there other R packages that accomplish the same thing? If so, how does yours differ or meet [our criteria for best-in-category](https://ropensci.github.io/dev_guide/policies.html#overlap)? No, there is not

-   (If applicable) Does your package comply with our [guidance around _Ethics, Data Privacy and Human Subjects Research_](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)? Yes it does

- [ x] I have read the [guide for authors](https://devguide.ropensci.org/guide-for-authors.html) and [rOpenSci packaging guide](https://devguide.ropensci.org/building.html).

- [ x] does not violate the Terms of Service of any service it interacts with.
- [x ] has a CRAN and OSI accepted license.
- [ x] contains a [README with instructions for installing the development version](https://ropensci.github.io/dev_guide/building.html#readme).
- [x ] includes [documentation with examples for all functions, created with roxygen2](https://ropensci.github.io/dev_guide/building.html#documentation).
- [ x] contains a vignette with examples of its essential functions and uses.
- [ x] has a [test suite](https://ropensci.github.io/dev_guide/building.html#testing).
- [ x] has [continuous integration](https://ropensci.github.io/dev_guide/ci.html), including reporting of test coverage using services such as Travis CI, Coveralls and/or CodeCov.

- [ x] Do you intend for this package to go on CRAN?

- [ x] The package is novel and will be of interest to the broad readership of the journal.
- [ x] The manuscript describing the package is no longer than 3000 words.
- [ x] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal (see [MEE's Policy on Publishing Code](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/journal-resources/policy-on-publishing-code.html))

- [ x] I agree to abide by [rOpenSci's Code of Conduct](https://ropensci.github.io/dev_guide/policies.html#code-of-conduct) during the review process and in maintaining my package should it be accepted.
