### 2021-06-02 #443 @haozhu233
### Editor checks:

- [X] **Automated tests:** Package has a testing suite and is tested via a CI service.

---

#### Editor comments

Hi @makinin, I'm your guest editor for the rdbhapi submission. I'm very sorry for the delay! I got dragged by many other things in my life. I'll help you move forward with the reviewing process as fast as I can.

I took a look at the package and I like its overall quality so far. I think the topic of this package is a good fit for ropensci and the existence of this package does make utilizing the DBH-API a lot easier (especially when I can't fully understand the Norwegian API documentation 😂). In terms of coding, I especially like how you use R's environment to host all the "private fields" of the API class if there is one. There are some small issues I found. If you can address them while I found reviewers for you, it would be great and it would also help the reviewers finish their reviewing process.

- It seems that this package is not on CRAN yet. The installation instruction in README is a little misleading.
- You do have a pkgdown website hosted at [https://makinin.github.io/rdbhapi/](https://makinin.github.io/rdbhapi/). Could you post this link in the link field of the github repository description? It will make it easier for people to find that site.
- After I run `goodpractice::gp()`, I got quite a few comments on code coverage and long code lines. If you can address them that will be great!
- There is a codecov setup for this repository. However, it seems that it has not been updated since 2 years ago. Is it possible for you to relink the code coverage check and the repository and update the code coverage number?

I will start to search for reviewers and I will keep you posted.

---

Reviewers:
Due date:
