### 2021-06-07 #443 @makinin
@haozhu233
Hello and thank you for your suggestions.
I have fixed long code lines and run checks after that but I get the following error message Project file "rdbhapi.Rproj" is not a valid MSBuild project file which must be an XML in appveyor and do not know how to fix it. I have run codecov again but since it was a lot of chaneges since first version for two years ago I have to write new examples.
Best regards,
Marija
