### 2021-06-07 #443 @haozhu233
No problem @makinin take your time! :)

For the appveyor issue, have you tried out https://github.com/krlmlr/r-appveyor? That might be helpful in this case.
