### 2021-12-02 #443 @haozhu233
@makinin Hi Marija, sorry for the long delay. For some reason I didn't see your reply.

In terms of the coverage issue, I noticed that the codecov.yml was changed since the last working check. Can you try to put back this removed line and see how it goes?

https://github.com/makinin/rdbhapi/blob/4dfe1f52b26c69c381b851df9290d8afb990a84c/codecov.yml#L14
